create store in separate module to help with fast refresh
