+++ src/components/exercise-view/state/exercise-view-store.ts
@@ -1,8 +1,3 @@
-import { exercisesData } from '@/content/exercises'
-import { generateData } from '@/data/generate-data'
-import { generateSeed } from '@/data/generate-seed'
-import { constrainedGeneration } from '@/helper/constrained-generation'
-import { isDeepEqual } from '@/helper/is-deep-equal'
 import { Store } from 'pullstate'
 
 export type IExerciseViewStore = {
@@ -15,6 +10,7 @@ export type IExerciseViewStore = {
   chatOverlay: null | 'solution'
 }
 
+console.log('create store')
 export const ExerciseViewStore = new Store<IExerciseViewStore>({
   id: -1,
   seed: '',
@@ -24,31 +20,3 @@ export const ExerciseViewStore = new Store<IExerciseViewStore>({
   navIndicatorExternalUpdate: -1,
   chatOverlay: null,
 })
-
-export function setupExercise(id: number) {
-  const content = exercisesData[id]
-  ExerciseViewStore.update(s => {
-    s.id = id
-    s.seed = generateSeed()
-    s.data = generateData(id, s.seed, content, true) as object
-    s.navIndicatorLength = 'tasks' in content ? content.tasks.length : 0
-    s.navIndicatorPosition = 0
-    s.navIndicatorExternalUpdate = 0
-  })
-}
-
-export function reseed() {
-  const s = ExerciseViewStore.getRawState()
-  const currentData = generateData(s.id, s.seed, exercisesData[s.id])
-  const newSeed = constrainedGeneration(
-    () => generateSeed(),
-    seed => {
-      const newData = generateData(s.id, seed, exercisesData[s.id])
-      return !isDeepEqual(currentData, newData)
-    },
-  )
-  ExerciseViewStore.update(s => {
-    s.seed = newSeed
-    s.data = generateData(s.id, newSeed, exercisesData[s.id]) as object
-  })
-}
